LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/pseudo-computed-style-stays-in-sync-with-new-element-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: incorrect viewbox undefined expected (string) "none" but got (undefined) undefined

Harness Error (FAIL), message = Unhandled rejection: assert_equals: incorrect viewbox undefined expected (string) "none" but got (undefined) undefined

TIMEOUT computed style on pseudo-element stays in sync with the DOM element Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_equals: incorrect viewbox undefined expected (string) "none" but got (undefined) undefined

TIMEOUT computed style on pseudo-element stays in sync with the DOM element Test timed out
